==> sandbox/01_collectArtists1.ipynb <==
## Config

# Import dependencies
import pandas as pd
import re
import os

from webdriver_manager.chrome import ChromeDriverManager

from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
from selenium.webdriver.support.ui import WebDriverWait
from selenium.common.exceptions import TimeoutException

from bs4 import BeautifulSoup

# Selenium/Chrome settings
chrome_options = webdriver.ChromeOptions()
#chrome_options.binary_location = os.environ.get("GOOGLE_CHROME_BIN")
#chrome_options.add_argument("--headless")
chrome_options.add_argument("--disable-dev-shm-usage")
chrome_options.add_argument("--no-sandbox")
capa = DesiredCapabilities.CHROME
capa["pageLoadStrategy"] = "none"
driver = webdriver.Chrome(
    executable_path=os.environ.get("CHROMEDRIVER_PATH"), 
    chrome_options=chrome_options, 
    desired_capabilities=capa)

## Scrape

# Get the html with Selenium
driver.get('https://www.musicfestivalwizard.com/all-festivals/')
# Select the html tag that contains the list of artists
x = driver.find_element(By.ID, value="artist").get_attribute("outerHTML")
# Create a BeautifulSoup object of the html
soup = BeautifulSoup(x)
# Get the artist's names in list form
artists = [a.text for a in soup.find_all("option")][1:]
# Quit the Selenium session
driver.quit()
# Convert to dataframe
result = pd.DataFrame()
result['artistName'] = artists
result

## Write to csv

result.to_csv('newArtistIdeas.csv', index=False)
