==> sandbox/02_artistsToSongs1.ipynb <==
## Config

# Import dependencies
import tekore as tk
import os
import time
import pandas as pd
import numpy as np
from datetime import datetime
from dateutil.relativedelta import relativedelta

# Environment variables
CLIENTID = os.environ.get('CLIENTID')
CLIENTSECRET = os.environ.get('CLIENTSECRET')
# Set global variables
MINDATE = datetime((datetime.today() - relativedelta(years=5)).year, 3, 17)


### Authenticate Tekore
# Get client token
app_token = tk.request_client_token(CLIENTID, CLIENTSECRET)
# Create spotify instance
spotify = tk.Spotify(app_token)

## Read in list of artists

# Read in list of artists
artists = pd.read_csv('artists1.csv')
artists

## Get artist IDs

# Create function to search spotify based on artist name
def searchSpotify(artists):
    try:
        # Search spotify for artists based on name
        possibleArtists, = spotify.search(artists['artistName'], types=['artist'], limit=1)
        # Select the first search result
        artists['artistID'] = possibleArtists.items[0].id
        artists['artistPop'] = possibleArtists.items[0].popularity
        artists['artistGenre'] = ", ".join(list(possibleArtists.items[0].genres))
        result = artists
    except: 
        artists['artistID'] = np.nan
        artists['artistPop'] = np.nan
        artists['artistGenre'] = np.nan
        result = artists
    finally:
        # Wait 2 seconds so we do not exceed our API call allowance
        time.sleep(2)
        return result

# Find an artistID for each artist based on artistName
artists = artists.apply(searchSpotify, axis=1)
artists

# Sort by popularity
artists = artists.sort_values(by='artistPop', ignore_index=True, ascending=False)
# Drop na values
artists = artists.dropna(subset='artistID')
artists

## Get top songs for each

# Define function that will get the top tracks for each artist in the list
def fetchTopSongs(row):
    try:
        # Query the spotify API
        topSongs = spotify.artist_top_tracks(row['artistID'], "US")
        # Iterate over the tracks to find a list of songURIs
        # Limit this to only those songs produced in the last few years; Use only the year (first four digits)
        # Limit this to only three songs per artist
        result = [x.uri for x in topSongs if datetime.strptime(x.album.release_date[:4], '%Y')>MINDATE][:3]
    except: 
        result = [np.nan]
    finally:
        # Wait 2 seconds so we do not exceed our API call allowance
        time.sleep(2)
        return result

# Drop artists who could not be found in the search 
songs1 = artists.copy()
# Find a list of songIDs for each artist
songs1['songURI'] = songs1.apply(fetchTopSongs, axis=1)
# Transform lists of songIDs to multiple rows
songs1 = songs1.explode('songURI', ignore_index=True)
# Drop where songs could not be found
songs1 = songs1.dropna()
songs1

## Load

songs1.to_csv('songs1.csv', index=False)
