==> sandbox/04_artistsToSongs2.ipynb <==
## Config

# Import dependencies
import tekore as tk
import os
import time
import pandas as pd
import numpy as np
from datetime import datetime, date
from dateutil.relativedelta import relativedelta

# Environment variables
CLIENTID = os.environ.get('CLIENTID')
CLIENTSECRET = os.environ.get('CLIENTSECRET')
REDIRECT_URI = 'https://example.com/callback'
# Set global variables
MINDATE = datetime((date.today() - relativedelta(years=3)).year, 6, 1)

### Authenticate Tekore and Authorize App
# Get client token
app_token = tk.request_client_token(CLIENTID, CLIENTSECRET)
# Get a user token; Note the need to ask for private read scope in order to read private playlists
# This will open a browser window; the user will need to copy the URL from the browser and paste it into the VSCode Command Palette
user_token = tk.prompt_for_user_token(CLIENTID, CLIENTSECRET, REDIRECT_URI, scope=tk.scope.playlist_read_private)
# Create spotify instance
spotify = tk.Spotify(app_token)
# Get the user's spotifyID; The spotify instance will have to reference the user_token to be authorized for this data
with spotify.token_as(user_token):
    userID = spotify.current_user().id

## Get user playlists

# Create a playlist for the user
with spotify.token_as(user_token):
    playlists = spotify.playlists(userID, limit=40)
    allPlaylists = [{'id':x.id, 'name':x.name} for x in playlists.items]

# I know which playlists I want to pull from
priorSummerNames = [
    'Summer 2018',
    'Summer 2019',
    'Summer 2020',
    'Summer 2021',
    'Summer 2022',
    '2023',
    'anti pop',
    'big on the internet',
    'Lorem',
    'Alternative Hip-Hop',
    'Et-Alt'
]
# Get the IDs of those playlists
priorSummerIDs = [{"playlistID":x.id, "artistSource":x.name} for x in playlists.items if x.name in priorSummerNames]
priorSummerIDs = pd.DataFrame(priorSummerIDs)

## Get artists in those playlists

# Get all the artistIDs in those playlists
# Start with an empty list
artists = []
# Loop through the playlists
for playlistID in priorSummerIDs['playlistID']:
    # Get the artistIDs
    nestedList = [[{"playlistID":playlistID, "artistID":y.id, "artistName":y.name} for y in x.track.artists] for x in spotify.playlist_items(playlistID).items]
    # Unnest the list
    flatList = [item for sublist in nestedList for item in sublist]
    # Append to the empty list
    artists += flatList
# Get unique artists
artists = pd.DataFrame(artists)
artists = artists.merge(priorSummerIDs, how='left', left_on='playlistID', right_on='playlistID').drop(columns='playlistID')
# Create a categorical sort order
artists['artistSource'] = pd.Categorical(artists['artistSource'], categories=priorSummerNames)
artists = artists.sort_values(by='artistSource', ignore_index=True).drop_duplicates(subset='artistID', ignore_index=True)

## Get artist genres

# Cut the df into chunks
n = 0
artistInfo = []
while n < len(artists):
    toGet = artists.iloc[n:(n+50)]
    # Add songs to the playlist
    with spotify.token_as(user_token):
        additions = [{'artistID':x.id, 'artistPopularity':x.popularity, 'artistGenre':x.genres} for x in spotify.artists(list(toGet['artistID']))]
    n += 50
    # Wait 2 seconds so we do not exceed our API call allowance
    time.sleep(2)
    artistInfo += additions
artistInfo = pd.DataFrame(artistInfo)
# Select only the first genre in the list
artistInfo['artistGenreUnique'] = [x[0] if len(x)>0 else 0 for x in artistInfo['artistGenre']]
# Merge song IDs in with artists and albums
artists = artists.merge(artistInfo, how='left', left_on='artistID', right_on='artistID')

## Get albums for each artist

# Define function that will get the top tracks for each artist in the list
def fetchRecentAlbums(artistID):
    try:
        # Query the spotify API
        query = spotify.artist_albums(artistID, limit=50, include_groups=['album', 'single']).items
        result = [{'artistID':artistID, 'albumID':x.id, 'albumURI':x.uri, 'albumName':x.name, 'albumReleaseDate':x.release_date, 'albumType':x.album_type} for x in query if datetime.strptime(x.release_date[:4], '%Y')>=MINDATE]
    except: 
        # If exception, fill in blank data
        pass
    finally:
        # Wait 2 seconds so we do not exceed our API call allowance
        time.sleep(2)
    return result

# Create empty list
albums = []
# Loop through list of artists
for artistID in artists['artistID']:
    # Get all recent albums by that artist
    additions = fetchRecentAlbums(artistID)
    # Append the recent albums to the list
    albums += additions
# Convert to new df
albums = pd.DataFrame(albums)

# Merge the albums into the artists
artists = artists.merge(albums, how='left', left_on='artistID', right_on='artistID')
# Drop any duplicates
artists = artists.drop_duplicates(subset=['artistID', 'albumName'])
artists = artists.dropna(subset=['albumID'])

## Get track URIs

artists

spotify.album('4jJpB6oqm4IiyyUYqetD7i', market='from_token')

# Cut the df into chunks
albumsUnique = artists.drop_duplicates(subset='albumID', ignore_index=True)
n = 0
albums = []
while n < len(artists):
    albumsToAdd = albumsUnique.iloc[n:(n+20)]
    # Add songs to the playlist
    try:
        with spotify.token_as(user_token):
            additions = [{'albumID':x.id, 'albumPopularity':x.popularity, 'genre':x.genres, 'songURI':[y.uri for y in x.tracks.items]} for x in spotify.albums(list(albumsToAdd['albumID']), market='from_token')]
            albums += additions
            # Wait 2 seconds so we do not exceed our API call allowance
            time.sleep(2)
    except:
        with spotify.token_as(user_token):
            for albumID in albumsToAdd['albumID']:
                try:
                    addition = [{'albumID':x.id, 'albumPopularity':x.popularity, 'genre':x.genres, 'songURI':[y.uri for y in x.tracks.items]} for x in spotify.album(albumID, market='from_token')]
                    albums += addition
                    # Wait 2 seconds so we do not exceed our API call allowance
                    time.sleep(2)
                except: 
                    pass
    n += 20
albums = pd.DataFrame(albums).explode('songURI')

# Merge song IDs in with artists and albums
songs = artists.merge(albums, how='left', left_on='albumID', right_on='albumID')
# Drop any duplicates
songs = songs.drop_duplicates(subset='songURI', ignore_index=True)

## Clean up columns

songs['artistGenre'] = [[y for y in x] for x in songs['artistGenre']]

songs.info()

## Write to csv

songs.to_csv('songs2.csv', index=False)
