==> sandbox/05_songsToPlaylist2.ipynb <==
## Config

# Import dependencies
import tekore as tk
import os
import time
import pandas as pd
import numpy as np
from datetime import datetime, date
from dateutil.relativedelta import relativedelta

# Environment variables
CLIENTID = os.environ.get('CLIENTID')
CLIENTSECRET = os.environ.get('CLIENTSECRET')
REDIRECT_URI = 'https://example.com/callback'

### Authenticate Tekore and Authorize App
# Get client token
app_token = tk.request_client_token(CLIENTID, CLIENTSECRET)
# Get a user token; Note the need to ask for write scope in order to create/edit playlists
# This will open a browser window; the user will need to copy the URL from the browser and paste it into the VSCode Command Palette
user_token = tk.prompt_for_user_token(CLIENTID, CLIENTSECRET, REDIRECT_URI, scope=tk.scope.write)
# Create spotify instance
spotify = tk.Spotify(app_token)
# Get the user's spotifyID; The spotify instance will have to reference the user_token to be authorized for this data
with spotify.token_as(user_token):
    userID = spotify.current_user().id

## Get some songs and sort them

# Extract list of songs
songs = pd.read_csv('songs2.csv')

# Clean
# Convert to datetime
songs['albumReleaseDate'] = pd.to_datetime(songs['albumReleaseDate'])

# Sort
# Create a categorical sort order
priorPlaylists = [
    '2023',
    'Summer 2022',
    'Summer 2021',
    'Summer 2020',
    'Summer 2019',
    'Summer 2018',
    'anti pop',
    'big on the internet',
    'Lorem',
    'Alternative Hip-Hop',
    'Et-Alt'
]
songs['artistSource'] = pd.Categorical(songs['artistSource'], categories=priorPlaylists)
songs = songs.sort_values(by=['artistSource', 'artistGenreUnique', 'artistID', 'albumID'], ignore_index=True)

# Filter
# Date
MINDATE = datetime((date.today() - relativedelta(years=1)).year, 6, 1)
filDate = songs.albumReleaseDate>=MINDATE
# Album Type
filAType = songs.albumType=='album'
# Filter
songs = songs.loc[filDate & filAType].reset_index(drop=True)

songsA = songs

# Filter
# Date
MINDATE = datetime((date.today() - relativedelta(years=1)).year, 3, 17)
MAXDATE = datetime((date.today() - relativedelta(years=1)).year, 5, 31)
filDateMin = songs.albumReleaseDate>=MINDATE
filDateMax = songs.albumReleaseDate<=MAXDATE
# Album Type
filAType = songs.albumType=='album'
# Playlist type
myPlaylists = [
    '2023',
    'Summer 2022',
    'Summer 2021',
    'Summer 2020',
    'Summer 2019',
    'Summer 2018',
]
filPlay = songs.artistSource.isin(myPlaylists)
# Filter
songs = songs.loc[filDateMin & filDateMax & filAType & filPlay].reset_index(drop=True)

songs

# Select the next chunk
# Number of entries
lastNumber = 2000
newNumber = lastNumber+4000
songs = songs.iloc[lastNumber:newNumber]

# Sort by Popularity from center
# songs['sorter'] = songs['artistPopularity'] - 50
# songs['sorter'] = songs['sorter'].abs()
# songs = songs.sort_values('sorter', ignore_index=True)

## Create an empty playlist in the user's account

# Create a playlist for the user
with spotify.token_as(user_token):
    newPlaylist = spotify.playlist_create(userID, "Summer Songs 2", public=False, description='Recent albums from Artist Candidates')
playlistID = newPlaylist.id

# Clear an existing playlist
with spotify.token_as(user_token):
    spotify.playlist_clear(playlistID)

## Add the songs to the playlist

# Cut the df into chunks
n = 0
while n < len(songs):
    songsToAdd = songs.iloc[n:(n+100)]
    # Add songs to the playlist
    try:
        with spotify.token_as(user_token):
            additions = spotify.playlist_add(playlistID, list(songsToAdd['songURI']), position=None)
            # Wait 2 seconds so we do not exceed our API call allowance
            time.sleep(2)
    except: 
        with spotify.token_as(user_token):
            for songURI in songsToAdd['songURI']:
                try:
                    additions = spotify.playlist_add(playlistID, songURI, position=None)
                    # Wait 2 seconds so we do not exceed our API call allowance
                    time.sleep(2)
                except: 
                    pass
    n += 100

songs.loc[songs['artistName']=='Justine Skye']

songs.iloc[6000]
